报告下载完成

+++ src/component/layout/layout.tsx
@@ -1,8 +1,8 @@
 import localforage from 'localforage';
 import { useNavigate } from 'react-router-dom';
 import { FC, PropsWithChildren, MouseEvent } from 'react';
-import { SettingOutlined, LogoutOutlined } from '@ant-design/icons';
-import { Button, message } from 'antd';
+import { LogoutOutlined } from '@ant-design/icons';
+import { App, Button, message } from 'antd';
 import { useModel } from '@/model';
 import { closeSse } from '@/utility/sse';
 import { StorageKeys } from '@/utility/storage-keys';
@@ -15,6 +15,7 @@ import { SettingMenuAction } from '../setting-menu/prop';
 const Layout: FC<PropsWithChildren<{}>> = ({ children }) => {
 
     const navigator = useNavigate();
+    const { modal } = App.useApp();
     const {
         loginUserName,
         logout,
@@ -23,20 +24,35 @@ const Layout: FC<PropsWithChildren<{}>> = ({ children }) => {
         logout: state.logout
     }));
 
-    const onLogoutClick = async (event: MouseEvent) => {
+    /**
+     * 登出Click
+     */
+    const onLogoutClick = (event: MouseEvent) => {
         event.preventDefault();
-        try {
-            closeSse();
-            await logout();
-            await localforage.clear();
-            sessionStorage.clear();
-            message.success(`用户${loginUserName}已登出`);
-            navigator('/');
-        } catch (error) {
-            console.warn(error);
-        }
+        modal.confirm({
+            onOk: async () => {
+                try {
+                    closeSse();
+                    await logout();
+                    await localforage.clear();
+                    sessionStorage.clear();
+                    message.success(`用户${loginUserName}已登出`);
+                    navigator('/');
+                } catch (error) {
+                    console.warn(error);
+                }
+            },
+            centered: true,
+            cancelText: '否',
+            okText: '是',
+            content: '确认登出系统？',
+            title: '登出'
+        });
     };
 
+    /**
+     * 菜单项Click
+     */
     const onMenuAction = (type: SettingMenuAction) => {
         switch (type) {
             case SettingMenuAction.Device:
@@ -57,16 +73,10 @@ const Layout: FC<PropsWithChildren<{}>> = ({ children }) => {
         <Reading />
         <div className="banner">
             <div>
-                用户: {sessionStorage.getItem(StorageKeys.User)}
+                用户: {sessionStorage.getItem(StorageKeys.User) ?? '-'}
             </div>
             <div>
                 <SettingMenu onMenuAction={onMenuAction} />
-                {/* <Button
-                onClick={() => navigator('/device')}
-                type="primary">
-                <SettingOutlined />
-                <span>设备管理</span>
-                </Button> */}
                 <Button
                     onClick={onLogoutClick}
                     type="primary"
